fix: Use existing TopBar for memory display instead of creating new bar
- Memory info now concatenates to the mainWindowTitle
- Shows as 'Title • 450MB • LOW' when enabled
- Removed unnecessary WindowDragBar component
- Simplified implementation to just append text
